rewrite tests without kaggle
